AP-8340: Introduce DLPythonKernelService on Python side
From now on, this is the central class for sending messages/issuing requests to Java. Requests are sent synchronously to avoid race conditions/enforce compliance with the Python kernel's messaging protocol.
The service class also allows client classes (e.g. DLPythonNetwork) access to the workspace of the current kernel execution.

We also introduce a new base class for network data suppliers, DLPythonNetworkInputBatchGenerator, which contains functionality that was previously part of DLKerasNetwork.DLDataSupplier. DLKerasNetwork.DLDataSupplier will be replaced by a Keras specific implementation of this base class.