AP-8036 : Follow up: Fix bug where optionals are printed
This is a follow up for the Keras learner node. See a previous commit for the executor node.

We also restructure the code (in executor and learner) a bit as the calculation of the converter's output size is only needed if there's an input size to compare to.
We also move the saving of settings to the end of the method (executor). Makes no sense to have it before the actual validation.
Also some code formatting (executor) and renaming (executor, learner).