AP-8454: Redefine DLViewData
Previously, there was a 1:1 relationship between view spec and view data. Now there is a 1:n relationship. "DLViewData" in its old sense is replaced by "DLViewDataCollection". View data now does not represent the data of an entire view but just of parts of it (e.g. a single line in a line plot).