AP-8454: Refactor Keras learner node view
The need for dynamically created view specs (based on whether validation data is present) introduced additional state in the node model and node view.
This commit mitigates the added complexity for the view by separating it into the node view itself which manages its lifecycle and the actual content panel which now maintains less state.